refactor(web): extract TimelineEntryList from ItemTimeline, no behaviour change (IDEA-2843)
Comments move under the item content on Details while Activity keeps
changes and Versions keeps versions, so the one feed has to render in two
DOM locations. One component instance cannot be in two places, and the
constraint on this work is ONE subscription and ONE composer — so the
rendered list becomes its own presentational component and ItemTimeline
stays the single owner of fetching, SSE, pagination, the attachment probe,
the paint fence and every mutation.

This commit is the extraction only. Nothing moves location and no
behaviour changes; the second mount site is the next commit.

- TimelineEntryList.svelte: the entry loop, rail chrome, the five card
  branches and their CSS, lifted verbatim. `listEl` is bindable because
  the owner's delegated lightbox listeners and imperative image-a11y pass
  attach to the container and stay with the owner. `showEmpty` is passed
  rather than derived from the rendered entries, preserving the owner's
  condition over the WHOLE feed — a tab that filters everything out must
  render an empty list, not claim the item has no history.
- The comment-card callbacks are optional here with no-op defaults: a list
  rendering no comments has nothing to hand them, and a card that could
  call one only renders when the owner supplied the real handler.

Evidence, and the reason it counts: the existing suite passes UNCHANGED —
116 files, 1983 tests — and svelte-check reports 0 errors. A refactor
whose only claim is "nothing changed" is exactly where an untouched suite
is the right instrument, but only if it actually exercises the moved
markup. It does: rendering the list over an empty array instead of
`entries` fails 64 tests across 7 files.
